refactor: move submissions API to new folder
